# app/public/docs/api.yaml
openapi: "3.0.3"
info:
  title: MagicWebb API
  version: "1.0.0"
  description: |
    MagicWebb is a decentralized NFT marketplace on Flare Coston2.
    All `/api/v1/*` routes are rate-limited (60 req/min per IP).
    Routes marked 🔒 require a valid SIWE JWT (Bearer token or `mw_s_*` cookie).
servers:
  - url: https://magicwebb.fly.dev
    description: Production (Flare Coston2 testnet)
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/v1/listings:
    get:
      summary: List active listings
      description: |
        Returns active, non-orphaned, non-expired listings.
        Filterable by collection and seller. Results include denormalized
        token name, image URI, and collection verified status.
      tags: [Listings]
      parameters:
        - name: collection
          in: query
          schema: { type: string, example: "0x1e9ab57e6147ad7a6c2aa7f64d61bcb10ca12a7c" }
          description: Filter by collection address
        - name: seller
          in: query
          schema: { type: string, example: "0xabcd...1234" }
          description: Filter by seller address
        - name: sort
          in: query
          schema: { type: string, enum: [recent, price_asc, price_desc, ending], default: recent }
          description: Sort order; `ending` = soonest expiry first
        - name: min_price
          in: query
          schema: { type: string, example: "1000000000000000000" }
          description: Minimum price in wei (inclusive)
        - name: max_price
          in: query
          schema: { type: string, example: "5000000000000000000" }
          description: Maximum price in wei (inclusive)
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
          description: 1-based page; pages beyond the first add OFFSET (page-1)*limit
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
          description: Max results (clamped to 1-200)
        - name: traits
          in: query
          schema: { type: string, example: "Background=Blue,Eyes=Laser" }
          description: Comma-separated trait_type=value pairs (AND across traits)
      responses:
        "200":
          description: Array of listing rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ListingRow"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/listings/{collection}/{id}:
    get:
      summary: Get single listing
      tags: [Listings]
      parameters:
        - name: collection
          in: path
          required: true
          schema: { type: string, example: "0x..." }
        - name: id
          in: path
          required: true
          schema: { type: string, example: "1" }
          description: Token ID (decimal)
      responses:
        "200":
          description: Listing row
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingRow"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/listings/{collection}/{id}/preflight:
    get:
      summary: Listing preflight check
      description: |
        Verifies fillability and repairs stale nft_ownership rows via
        on-chain RPC call when the DB projection is stale.
      tags: [Listings]
      parameters:
        - name: collection
          in: path
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: seller
          in: query
          required: true
          schema: { type: string, example: "0x..." }
          description: Seller address to verify
      responses:
        "200":
          description: Preflight result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:          { type: boolean, description: "Listing is fillable" }
                  listed:      { type: boolean }
                  orphaned:    { type: boolean }
                  seller_owns: { type: boolean }
                  price_wei:   { type: string, description: "Price in wei (decimal string)" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Auctions ────────────────────────────────────────────────────────
  /api/v1/token/{collection}/{id}:
    get:
      summary: Get indexed token detail
      description: |
        Everything the indexer knows about one token: metadata, current
        owner, last sale and when it was indexed, plus the collection's card
        facts. A token the indexer has never seen is a 404 — never an empty
        row — so the token page can say "not indexed yet".
      tags: [Listings]
      parameters:
        - name: collection
          in: path
          required: true
          schema: { type: string, example: "0x..." }
        - name: id
          in: path
          required: true
          schema: { type: string, example: "1" }
          description: Token ID (decimal)
      responses:
        "200":
          description: Token detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: "SIWE JWT via Authorization: Bearer <token>"
    CookieAuth:
      type: apiKey
      in: cookie
      name: mw_s_
      description: "HttpOnly session cookie with addr prefix (e.g. mw_s_0xabcd...)"
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, example: "internal error" }
  schemas:
...
    ListingRow:
      type: object
      properties:
        collection:            { type: string }
        token_id:              { type: string, description: "Token ID (decimal uint256)" }
        seller:                { type: string }
        price_wei:             { type: string }
        amount:                { type: integer }
        standard:              { type: string, enum: [erc721, erc1155] }
        expires_at:            { type: string, format: date-time }
        listed_at:             { type: string, format: date-time }
        tx_hash:               { type: string }
        name:                  { type: string }
        image_uri:             { type: string }
        total_supply:          { type: integer, description: "Collection-level total supply (0 when unindexed)" }
        collection_verified:   { type: boolean, description: "Collection passed the standard + metadata verifier" }
        collection_creator:    { type: string, description: "Collection ERC-173 owner(); empty when unresolved. Verified + creator = Authentic badge" }
        collection_name:       { type: string }
        collection_tracked:    { type: boolean, description: "True when the indexer tracks the collection (distinguishes unknown contract from known-but-unverified)" }

    # ── Auctions ──────────────────────────────────────────────────────
...
    TokenDetail:
      type: object
      properties:
        collection:          { type: string }
        token_id:            { type: string }
        name:                { type: string }
        description:         { type: string }
        image_uri:           { type: string }
        animation_uri:       { type: string }
        metadata_uri:        { type: string }
        owner:               { type: string, description: "Current holder; empty when unknown" }
        indexed_at:          { type: string, format: date-time }
        last_sale_wei:       { type: string, description: "Empty when never sold" }
        last_sale_at:        { type: string, format: date-time, nullable: true }
        collection_name:     { type: string }
        standard:            { type: string, enum: [erc721, erc1155, ""] }
        collection_verified: { type: boolean }
        collection_creator:  { type: string }
        collection_tracked:  { type: boolean }
...